@@ -2,7 +2,7 @@ import 'package:cr_logger/cr_logger.dart';
 import 'package:cr_logger/src/data/sqflite_db/entities/http_entity.dart';
 import 'package:cr_logger/src/utils/parsers/isolate_parser.dart';
 
-class HttpEntityConverter {
+final class HttpEntityConverter {
   final _parser = IsolateParser();
 
   Future<HttpBean> inToOut(HttpEntity inObject) async {
