@@ -2,7 +2,7 @@ import 'package:cr_logger/cr_logger.dart';
 import 'package:cr_logger/src/utils/enum_ext.dart';
 import 'package:cr_logger/src/utils/map_ext.dart';
 
-class LogEntity {
+final class LogEntity {
   LogEntity({
     required this.id,
     required this.message,
