@@ -2,7 +2,7 @@ import 'package:cr_logger/src/page/actions_and_values/models/notifier_data.dart'
 import 'package:flutter/cupertino.dart';
 
 /// Manager through which value notifiers are added to the page
-class NotifiersManager {
+final class NotifiersManager {
   NotifiersManager._();
 
   static final List<NotifierData> valueNotifiers = [];
