@@ -17,8 +17,8 @@ class WebHeaderWidget extends StatelessWidget {
     return Row(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: [
-        Row(
-          children: const [
+        const Row(
+          children: [
             SizedBox(width: 16),
             Text(
               'Logs',
