@@ -1,14 +1,15 @@
+import 'package:cr_logger/src/constants.dart';
 import 'package:cr_logger/src/controllers/logs_mode_controller.dart';
 import 'package:cr_logger/src/cr_logger_helper.dart';
 import 'package:cr_logger/src/data/bean/log_bean.dart';
 import 'package:cr_logger/src/data/models/log_type.dart';
 import 'package:cr_logger/src/extensions/extensions.dart';
+import 'package:cr_logger/src/page/widgets/json_details_widget.dart';
 import 'package:cr_logger/src/res/colors.dart';
 import 'package:cr_logger/src/res/styles.dart';
+import 'package:cr_logger/src/utils/text_with_params_widget.dart';
 import 'package:cr_logger/src/widget/copy_widget.dart';
 import 'package:cr_logger/src/widget/cr_app_bar.dart';
-import 'package:cr_logger/src/widget/expand_arrow_button.dart';
-import 'package:cr_logger/src/widget/json_widget/json_widget.dart';
 import 'package:cr_logger/src/widget/rounded_card.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -34,15 +35,7 @@ class LogLocalDetailPage extends StatefulWidget {
 }
 
 class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
-  static const logTypes = {
-    LogType.debug: 'Debug',
-    LogType.info: 'Info',
-    LogType.error: 'Error',
-    LogType.http: 'HTTP',
-  };
-
-  final listWidgetStackTrace = <Widget>[];
-  final _allExpandedNodesNotifier = ValueNotifier<bool>(true);
+  final _listWidgetStackTrace = <Widget>[];
 
   @override
   void initState() {
@@ -52,15 +45,17 @@ class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
 
   @override
   Widget build(BuildContext context) {
-    final logName = logTypes[widget.logType!];
+    final logType = widget.logType ?? LogType.debug;
+    final logName = logType.name;
     final log = widget.logBean;
-    final isJsonData = log?.message is Map<String, dynamic>;
+    final logMessage = log?.message;
+    final isJsonData = logMessage is Map<String, dynamic>;
 
     /// [!kIsWeb] because the logs may be imported
     /// and then there is no way to know the date of the logs.
     final time = LogsModeController.instance.isFromCurrentSession && !kIsWeb
-        ? log?.time.formatTime()
-        : log?.time.formatTimeWithYear();
+        ? log?.time.formatTime(context)
+        : log?.time.formatTimeWithYear(context);
 
     return Theme(
       data: CRLoggerHelper.instance.theme,
@@ -88,56 +83,26 @@ class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       /// Json data
-                      if (isJsonData) ...[
-                        ValueListenableBuilder(
-                          valueListenable: _allExpandedNodesNotifier,
-                          builder: (
-                            BuildContext context,
-                            bool isAllNodesExpanded,
-                            Widget? child,
-                          ) {
-                            return Column(children: [
-                              Row(
-                                mainAxisAlignment:
-                                    MainAxisAlignment.spaceBetween,
-                                children: [
-                                  /// Json title
-                                  const Text(
-                                    'JSON data',
-                                    style: CRStyle.subtitle1BlackSemiBold16,
-                                  ),
-
-                                  /// Json expand button
-                                  ExpandArrowButton(
-                                    isExpanded: isAllNodesExpanded,
-                                    onTap: _onExpandArrowTap,
-                                  ),
-                                ],
-                              ),
-
-                              /// Json widget
-                              JsonWidget(
-                                log.message as Map<String, dynamic>,
-                                allExpandedNodes: isAllNodesExpanded,
-                              ),
-                            ]);
-                          },
-                        ),
-                      ] else
+                      if (isJsonData)
+                        JsonDetailsWidget(logMessage: logMessage)
+                      else
 
                         /// Log message with copy widget
                         Row(
+                          crossAxisAlignment: CrossAxisAlignment.start,
                           mainAxisAlignment: MainAxisAlignment.spaceBetween,
                           children: [
                             Expanded(
-                              child: Text(
-                                'Message:\n${log.message.toString()}',
-                                style: CRStyle.bodyBlackMedium14.copyWith(
-                                  color: log.color,
-                                ),
+                              child: TextWithParamsWidget(
+                                logMessage,
+                                textColor: log.color,
+                                onParamTap: _onCopy,
                               ),
                             ),
-                            CopyWidget(onCopy: _onCopy),
+                            const SizedBox(width: 10),
+                            CopyWidget(
+                              onCopy: () => _onCopy(logMessage.toString()),
+                            ),
                           ],
                         ),
                       const SizedBox(height: 8),
@@ -150,8 +115,8 @@ class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
                       const SizedBox(height: 10),
 
                       /// Stacktrace
-                      if (listWidgetStackTrace.isNotEmpty)
-                        ...listWidgetStackTrace,
+                      if (_listWidgetStackTrace.isNotEmpty)
+                        ..._listWidgetStackTrace,
                     ],
                   ),
                 ),
@@ -170,8 +135,8 @@ class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
   /// to highlight the stack trace
   Future<void> getStackTraceListWidget() async {
     final packageInfo = await PackageInfo.fromPlatform();
-    if (listWidgetStackTrace.isNotEmpty) {
-      listWidgetStackTrace.clear();
+    if (_listWidgetStackTrace.isNotEmpty) {
+      _listWidgetStackTrace.clear();
     }
     setState(() {
       if (widget.logBean?.stackTrace != null) {
@@ -181,7 +146,7 @@ class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
         for (final stackCall in listModelStackTrace) {
           if (stackCall.contains(packageName)) {
             final stackTraceParts = stackCall.split('package:');
-            listWidgetStackTrace.add(
+            _listWidgetStackTrace.add(
               RichText(
                 text: TextSpan(
                   children: [
@@ -205,7 +170,7 @@ class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
               ),
             );
           } else {
-            listWidgetStackTrace.add(
+            _listWidgetStackTrace.add(
               Text(
                 stackCall,
                 style: CRStyle.h3Black,
@@ -217,30 +182,19 @@ class _LogLocalDetailPageState extends State<LogLocalDetailPage> {
     });
   }
 
-  void _onCopy() {
-    final logMessage = widget.logBean?.message.toString();
+  void _onCopy(String text) {
+    final message = text.replaceAll(patternOfParamsRegex, '');
 
     ScaffoldMessenger.of(context).clearSnackBars();
     ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(
         content: Text(
-          'Copy \n"$logMessage"',
+          'Copy \n"$message"',
           maxLines: 4,
           overflow: TextOverflow.ellipsis,
         ),
       ),
     );
-
-    if (logMessage != null) {
-      Clipboard.setData(
-        ClipboardData(
-          text: logMessage,
-        ),
-      );
-    }
-  }
-
-  void _onExpandArrowTap() {
-    _allExpandedNodesNotifier.value = !_allExpandedNodesNotifier.value;
+    Clipboard.setData(ClipboardData(text: message));
   }
 }
