@@ -1,6 +1,6 @@
 import 'package:logger/logger.dart';
 
-class NothingLogFilter extends LogFilter {
+final class NothingLogFilter extends LogFilter {
   @override
   bool shouldLog(LogEvent event) {
     return false;
