@@ -1,4 +1,4 @@
-class Pair<A, B> {
+final class Pair<A, B> {
   Pair({
     required this.first,
     required this.second,
