@@ -3,22 +3,18 @@ import 'dart:convert';
 import 'package:worker_manager/worker_manager.dart';
 
 /// This class is for converting a [String] into a [Map] and conversely.
-class IsolateParser {
-  ///[data] is a string, but may contain a map, so a decoding attempt is made first. If it fails, [data] is returned as a string.
-  dynamic decode(String data) => Executor().execute(
-        arg1: data,
-        fun1: _decode,
-      );
+final class IsolateParser {
+  /// [data] is a string, but may contain a map, so a decoding attempt is made first. If it fails, [data] is returned as a string.
+  dynamic decode(String data) =>
+      workerManager.execute(() async => _decode(data));
 
   /// [data] can be any type, so there is a check if it is a map, then [json.encode] is called,
   /// otherwise [data] is converted into a string and returned.
-  Future<String> encode(dynamic data) => Executor().execute(
-        arg1: data,
-        fun1: _encode,
-      );
+  Future<String> encode(dynamic data) =>
+      workerManager.execute(() async => _encode(data));
 }
 
-dynamic _decode(String data, _) {
+dynamic _decode(String data) {
   try {
     return json.decode(data);
   } catch (_) {
@@ -26,5 +22,5 @@ dynamic _decode(String data, _) {
   }
 }
 
-String _encode(dynamic data, _) =>
+String _encode(dynamic data) =>
     data is Map<String, dynamic> ? json.encode(data) : data.toString();
