@@ -1,5 +1,5 @@
 import 'package:cr_logger/src/utils/web_utils.dart';
-import 'package:cr_logger/src/widget/adaptive_layout/layout_type_inherited.dart';
+import 'package:cr_logger/src/widget/adaptive_layout/layout_types.dart';
 import 'package:flutter/material.dart';
 
 class AdaptiveLayoutWidget extends StatefulWidget {
@@ -12,7 +12,7 @@ class AdaptiveLayoutWidget extends StatefulWidget {
 
   final Widget mobileLayoutWidget;
   final Widget webLayoutWidget;
-  final LayoutTypeChanged? onLayoutChange;
+  final ValueChanged<LayoutType>? onLayoutChange;
 
   @override
   _AdaptiveLayoutWidgetState createState() => _AdaptiveLayoutWidgetState();
@@ -33,12 +33,9 @@ class _AdaptiveLayoutWidgetState extends State<AdaptiveLayoutWidget> {
           widget.onLayoutChange?.call(_layoutType);
         }
 
-        return LayoutTypeProvider(
-          layoutType: _layoutType,
-          child: _layoutType.isWebLayout
-              ? widget.webLayoutWidget
-              : widget.mobileLayoutWidget,
-        );
+        return _layoutType.isWebLayout
+            ? widget.webLayoutWidget
+            : widget.mobileLayoutWidget;
       },
     );
   }
